+++ SPTSharp/Models/Eft/Launcher/MiniProfile.cs
@@ -12,8 +12,8 @@ namespace SPTSharp.Models.Eft.Launcher
         public string nickname { get; set; }
         public string side { get; set; }
         public int currlvl { get; set; }
-        public float currexp { get; set; }
-        public float prevexp { get; set; }
+        public int currexp { get; set; }
+        public int prevexp { get; set; }
         public int nextlvl { get; set; }
         public int maxlvl { get; set; } 
         public AkiData akiData { get; set; }
